Update const.py
Updated code with ChatGTP to mitigate depreciation of the following error message:

DATA_MEGABYTES was used from synology_dsm, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfInformation.MEGABYTES instead, please report it to the author of the 'synology_dsm' custom integration

DATA_RATE_KILOBYTES_PER_SECOND was used from synology_dsm, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfDataRate.KILOBYTES_PER_SECOND instead, please report it to the author of the 'synology_dsm' custom integration

DATA_TERABYTES was used from synology_dsm, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfInformation.TERABYTES instead, please report it to the author of the 'synology_dsm' custom integration

TEMP_CELSIUS was used from synology_dsm, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfTemperature.CELSIUS instead, please report it to the author of the 'synology_dsm' custom integration

DATA_GIGABYTES was used from synology_dsm, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfInformation.GIGABYTES instead, please report it to the author of the 'synology_dsm' custom integration